Initial commit
A working program which currently supports options for plotting the
values of sigma, plotting the eigenfunctions and giving the
coefficients.

The biggest drawback at the moment is that the spherical triangle is
hard coded together with some of the optimizations that only hold for
this triangle. These optimizations are skipping every other term in
the expansion and only working with half of the boundary.