Add handling of integer values for mu
When mu is an exact integer special care has to be taken. When the
value of mu0 is not exactly representable as a floating point it could
happen that mu = k*mu0 is an exact integer in exact arithmetic but not
in floating point arithmetic. To handle this we simply check if the
bound for mu contains an integer and if so sets it to exactly this
integer. This works since we do not need a proof of the correctness in
this case.