Factor out the main function from particular_solution
Preparing for having the source be mainly a library and the executable
code be either examples or tests. The executable part of
particular_solution is moved to the example particular_solution_main.

At the moment a lot of functionality has been removed, there is no
support for plotting sigma or the eigenfunction for example, but this
will be added later as separate functions.