Fix comparison of values to use mpfr_cmp instead of <
This slightly changes some of the test results since the computed
minimum is different. You would expect the results to be slightly
better since we are now correctly computing the place for the minimum.
However for some reason the results are for most of the tests slightly
worse, most likely this is just noise. We update the tests to take
this into account.