Change the order the angles are used in
The original intention was to have in the first parameterization the
first angle on the north pole and the second angle have y equal to
zero. It turns out this have not been the case and instead it was the
third angle that had y equal to zero. This have now been fixed.